PopUp_WebApp_Python/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
httpx
python-dotenv
websockets
